# hotel-management-ui/src/openApi/openApi.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "openApi specification - user19",
      "description": "OpenApi documentation for hotel management",
      "termsOfService": "terms of service",
      "contact": {
        "name": "user19",
        "url": "http://user19.com/website",
        "email": "contact@User19.com"
      },
      "license": {
        "name": "licence name",
        "url": "http://some-url.com"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:9090/api/v1",
        "description": "local Env"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
...
      "/users/get-user-bookings/{userId}": {
        "get": {
          "tags": [
            "users"
          ],
          "operationId": "getUserBookingHistory",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponseBookingDTO"
                  }
                }
              }
            }
          }
        }
      },
...
      "/users/get-by-id/{userId}": {
        "get": {
          "tags": [
            "users"
          ],
          "operationId": "getUserById",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/users/all": {
        "get": {
          "tags": [
            "users"
          ],
          "operationId": "getAllUsers",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 5
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponseUserDTO"
                  }
                }
              }
            }
          }
        }
      },
...
      "/users/delete-by-id/{userId}": {
        "delete": {
          "tags": [
            "users"
          ],
          "operationId": "deleteUserById",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "BookingDTO": {
          "type": "object",
          "properties": {
            "checkInDate": {
              "type": "string",
              "format": "date"
            },
            "checkOutDate": {
              "type": "string",
              "format": "date"
            },
            "numOfAdults": {
              "type": "integer",
              "format": "int32"
            },
            "numOfChildren": {
              "type": "integer",
              "format": "int32"
            },
            "totalNumOfGuest": {
              "type": "integer",
              "format": "int32"
            },
            "bookingConfigurationCode": {
              "type": "string"
            },
            "user": {
              "$ref": "#/components/schemas/UserDTO"
            },
            "room": {
              "$ref": "#/components/schemas/RoomDTO"
            },
            "id": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "Role": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string"
            },
            "createdDate": {
              "type": "string",
              "format": "date-time"
            },
            "lastModifiedDate": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "RoomDTO": {
          "required": [
            "roomDescription",
            "roomPrice",
            "roomType"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "roomType": {
              "type": "string"
            },
            "roomPhoto": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "byte"
              }
            },
            "roomDescription": {
              "type": "string"
            },
            "roomPrice": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "UserDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "email": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "role": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Role"
              }
            },
            "phoneNumber": {
              "type": "string"
            },
            "bookings": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/BookingDTO"
              }
            }
          }
        },
...
        "PageResponseBookingDTO": {
          "type": "object",
          "properties": {
            "content": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/BookingDTO"
              }
            },
            "number": {
              "type": "integer",
              "format": "int32"
            },
            "totalElements": {
              "type": "integer",
              "format": "int64"
            },
            "size": {
              "type": "integer",
              "format": "int32"
            },
            "totalPages": {
              "type": "integer",
              "format": "int32"
            },
            "first": {
              "type": "boolean"
            },
            "last": {
              "type": "boolean"
            }
          }
        },
        "PageResponseUserDTO": {
          "type": "object",
          "properties": {
            "content": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/UserDTO"
              }
            },
            "number": {
              "type": "integer",
              "format": "int32"
            },
            "totalElements": {
              "type": "integer",
              "format": "int64"
            },
            "size": {
              "type": "integer",
              "format": "int32"
            },
            "totalPages": {
              "type": "integer",
              "format": "int32"
            },
            "first": {
              "type": "boolean"
            },
            "last": {
              "type": "boolean"
            }
          }
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "description": "JWT authentication description",
          "in": "header",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }